# feat: improves escapeString's performance