fix #306 

```
name           old time/op  new time/op  delta
306/stdjson-8   314µs ± 0%   312µs ± 1%   -0.61%  (p=0.000 n=10+8)
306/go-json-8  3.36ms ± 1%  0.12ms ± 0%  -96.50%  (p=0.000 n=10+8)
```